refactor: improve the way adding script
