+++ pages/_app.js
@@ -1,7 +1,26 @@
 import '../styles/globals.css'
+import Script from "next/script";
 
-function MyApp({ Component, pageProps }) {
-  return <Component {...pageProps} />
+function MyApp({Component, pageProps}) {
+  return <>
+    <Script id="takw">
+      {
+        `
+          var Tawk_API = Tawk_API || {}, Tawk_LoadStart = new Date();
+          (function () {
+            var s1 = document.createElement("script"), s0 = document.getElementsByTagName("script")[0];
+            s1.async = true;
+            s1.src = 'https://embed.tawk.to/62fa110a54f06e12d88eaf2e/1gagd54dv';
+            s1.charset = 'UTF-8';
+            s1.setAttribute('crossorigin', '*');
+            s0.parentNode.insertBefore(s1, s0);
+          })();
+        `
+      }
+    </Script>
+    <Component {...pageProps} />
+  </>
+
 }
 
 export default MyApp
